requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.1
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1
django-celery-beat==2.1.0
django-filter==2.4.0
django-timezone-field==4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
kombu==5.0.2
Markdown==3.3.3
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
